Organize protocol-kit types and export from barrel
